Add maths flavour of markdown
This flavour of markdown is designed to better facilite the use of the
maths subset of latex, rendered with mathjax. While vanillia markdown
works fine with blocks delimited by $$, it will add redundent <p> tags.
This flavour fixes that, and brings a bit more customisation on how the
parser treats maths blocks.