Delete design examples and additional css styles
I Removed the additional css styles that beautified the parsed and
rendered md's. Further I removed the examples in /design of using this.
I made the mistake of forgetting this repository is about parsing
markdown, save your adventures into design for another repo :D.

I have also changed the --full output, st the mathjax cdn script tag is
only included when the flavour is set to maths. Its not exactly elegant
code, but it does the job :D.